OCPP (Şarj İstasyonu) ve CAN bus güvenlik senaryosu/requirements.txt
ocpp>=0.20.0
websockets>=12.0
aiohttp>=3.9.0
python-can>=4.3.0
cantools>=39.0.0

